+++ NEWS.md
@@ -11,6 +11,7 @@ Development
 - Set right referrer header for password reset page ([#15699](https://github.com/CartoDB/cartodb/pull/15699))
 - Fix navigation bar tests
 - Ignore update_timestamp function on migrations ([#15710](https://github.com/CartoDB/cartodb/pull/15710))
+- Remove faulty ACLs in User Mover ([#15717](https://github.com/CartoDB/cartodb/pull/15717))
 
 4.38.0 (2020-06-05)
 -------------------